+++ src/compositions/Questionnaire/Questionnaire/Questionnaire.css
@@ -1,0 +1,4 @@
+.Questionnaire {
+    max-width: 600px;
+    margin: auto;
+}
